ChatBotApp/requirements.txt
uvicorn
requests
numpy
fastapi
sentence-transformers
watchdog
openai